requirements.txt
streamlit>=1.30,<2.0
pandas>=2.1,<3.0
altair>=5.0,<6.0
python-dateutil>=2.8,<3.0
numpy>=1.26,<2.0
psycopg2-binary>=2.9
pyotp
qrcode
pillow
fpdf2
streamlit-lottie

